requirements.txt
discord>=2.2.2
loguru>=0.6.0
python_dotenv>=0.21.0